Added Denoising autoencoder (DNAE) example. Added self supervised network base class, used for pretraining unsupervised tasks (such as DNAE or Jigsaw puzzles solving). Added MNIST classifier example alongside DNAE pretraining for comparison with regular random initializatiton.

+++ neural_wrappers/readers/corrupter_reader.py
@@ -16,8 +16,8 @@ class CorrupterReader(DatasetReader):
 		noiseValues = np.random.choice([Min, Max], p=[0.5, 0.5], size=inputs.shape)
 
 		# mask = 0 => inputs * 1 + noiseValues * 0 (no change)
-		# mask = 1 => inputs * 0 + noiseValues * 1 (change value at that place)		
-		newInputs = inputs * (1 - mask) + noiseValues * mask
+		# mask = 1 => inputs * 0 + noiseValues * 1 (change value at that place)
+		newInputs = (inputs * (1 - mask) + noiseValues * mask).astype(inputs.dtype)
 		return newInputs
 
 	def iterate_once(self, type, miniBatchSize):
